explanation Remove erroneous TODOs.
These were for an abandoned idea of adding a list of classes to each rule.
Something like "naming", "style", "synthsim", "intent", "backcompat", and
such like is what I had in mind.
The idea is abandoned for now because I haven't come up with a clear
definition of exactly what a "class" should be.