Remove clone call
Remove clone call that is not needed. 

(Background: I have run svlint on a large code base. The total run time was 14 minutes. I tried profiling the code and noticed that the top 3 functions were memory allocations. I have not had the time to find out exactly where those memory allocations are coming from. This is my best shot at improving performance with the limited time that I have had.

For comparison the verible-verilog-lint runs in less than a second)