Add missing error message
If a definition was not found inside an included file then the error would be `Include { source: DefineNotFound("FOO") }` which would previously not print a message at all. This prints a specific message and adds a catch all for other cases.

This should probably be a recursive function that builds a series of "in file included from..." messages that you see from C compilers, but that is a bigger change and this at least avoids having no error message at all.